Fix some details on unpeering

+++ pkg/liqoctl/network/handler.go
@@ -16,7 +16,6 @@ package network
 
 import (
 	"context"
-	"fmt"
 	"time"
 
 	corev1 "k8s.io/api/core/v1"
@@ -36,9 +35,11 @@ type Options struct {
 	LocalFactory  *factory.Factory
 	RemoteFactory *factory.Factory
 
-	Timeout        time.Duration
-	Wait           bool
-	SkipValidation bool
+	Timeout         time.Duration
+	Wait            bool
+	SkipValidation  bool
+	Force           bool
+	remoteClusterId liqov1beta1.ClusterID
 
 	ServerGatewayType           string
 	ServerTemplateName          string
@@ -104,20 +105,33 @@ func (o *Options) RunReset(ctx context.Context) error {
 	return cluster2.DeleteConfiguration(ctx, cluster1.localClusterID, cluster2.localNetworkNamespace)
 }
 
-func (o *Options) RunResetLocalOnly(ctx context.Context) error {
-	fmt.Print("Dentro runResetLocalONly")
+// RunReset reset the liqo networking in local cluster.
+func (o *Options) RunResetLocalOnly(ctx context.Context, remoteClusterId liqov1beta1.ClusterID) error {
+	o.Force = true
+	o.remoteClusterId = remoteClusterId
+
 	ctx, cancel := context.WithTimeout(ctx, o.Timeout)
 	defer cancel()
 
-	// Inizializza solo il cluster locale
+	// Create and initialize local cluster
 	cluster, err := NewCluster(ctx, o.LocalFactory, nil, false)
 	if err != nil {
 		return err
 	}
-	fmt.Print("Dopo new cluster")
 
-	// Cancella la configurazione locale (senza tentare di contattare il cluster remoto)
-	return cluster.DeleteConfiguration(ctx, "", cluster.localNetworkNamespace)
+	cluster.remoteClusterID = remoteClusterId
+	if err := cluster.SetNamespaces(ctx, false); err != nil {
+		return err
+	}
+
+	// Run disconnect command to remove gateways.
+
+	if err := o.RunDisconnect(ctx, cluster, nil); err != nil {
+		return err
+	}
+
+	// Delete Configuration on local cluster
+	return cluster.DeleteConfiguration(ctx, remoteClusterId, cluster.localNetworkNamespace)
 }
 
 // RunConnect connect two clusters using liqo networking.
@@ -307,7 +321,7 @@ func (o *Options) RunDisconnect(ctx context.Context, cluster1, cluster2 *Cluster
 		}
 	}
 
-	if cluster2 == nil {
+	if cluster2 == nil && !o.Force {
 		// Create and initialize cluster 2.
 		cluster2, err = NewCluster(ctx, o.RemoteFactory, o.LocalFactory, false)
 		if err != nil {
@@ -315,23 +329,37 @@ func (o *Options) RunDisconnect(ctx context.Context, cluster1, cluster2 *Cluster
 		}
 	}
 
-	// Delete gateway client on cluster 1
-	if err := cluster1.DeleteGatewayClient(ctx, cluster2.localClusterID); err != nil {
-		return err
+	if o.Force {
+		// Delete gateway client on cluster 1
+		if err := cluster1.DeleteGatewayClient(ctx, o.remoteClusterId); err != nil {
+			return err
+		}
+	} else {
+		// Delete gateway client on cluster 1
+		if err := cluster1.DeleteGatewayClient(ctx, cluster2.localClusterID); err != nil {
+			return err
+		}
 	}
 
-	// Delete gateway client on cluster 2
-	if err := cluster2.DeleteGatewayClient(ctx, cluster1.localClusterID); err != nil {
-		return err
+	if !o.Force {
+		// Delete gateway client on cluster 2
+		if err := cluster2.DeleteGatewayClient(ctx, cluster1.localClusterID); err != nil {
+			return err
+		}
+	}
+	if !o.Force {
+		// Delete gateway server on cluster 1
+		if err := cluster1.DeleteGatewayServer(ctx, cluster2.localClusterID); err != nil {
+			return err
+		}
+
+		// Delete gateway server on cluster 2
+		if err := cluster2.DeleteGatewayServer(ctx, cluster1.localClusterID); err != nil {
+			return err
+		}
 	}
 
-	// Delete gateway server on cluster 1
-	if err := cluster1.DeleteGatewayServer(ctx, cluster2.localClusterID); err != nil {
-		return err
-	}
-
-	// Delete gateway server on cluster 2
-	return cluster2.DeleteGatewayServer(ctx, cluster1.localClusterID)
+	return nil
 }
 
 func (o *Options) initNetworkConfigs(ctx context.Context, cluster1, cluster2 *Cluster) error {
